+++ packages/react/src/components/progress/progress-circle-range.tsx
@@ -1,18 +1,20 @@
 import { mergeProps } from '@zag-js/react'
-import type { SVGProps } from 'react'
-import { type PolymorphicProps, ark } from '../factory'
+import { forwardRef } from 'react'
+import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { useProgressContext } from './use-progress-context'
 
 export interface ProgressCircleRangeBaseProps extends PolymorphicProps {}
 export interface ProgressCircleRangeProps
-  extends SVGProps<SVGCircleElement>,
+  extends HTMLProps<'circle'>,
     ProgressCircleRangeBaseProps {}
 
-export const ProgressCircleRange = (props: ProgressCircleRangeProps) => {
-  const progress = useProgressContext()
-  const mergedProps = mergeProps(progress.getCircleRangeProps(), props)
+export const ProgressCircleRange = forwardRef<SVGCircleElement, ProgressCircleRangeProps>(
+  (props, ref) => {
+    const progress = useProgressContext()
+    const mergedProps = mergeProps(progress.getCircleRangeProps(), props)
 
-  return <ark.circle {...mergedProps} />
-}
+    return <ark.circle ref={ref} {...mergedProps} />
+  },
+)
 
 ProgressCircleRange.displayName = 'ProgressCircleRange'