+++ packages/react/src/components/progress/progress-circle-track.tsx
@@ -1,18 +1,20 @@
 import { mergeProps } from '@zag-js/react'
-import type { SVGProps } from 'react'
-import { type PolymorphicProps, ark } from '../factory'
+import { forwardRef } from 'react'
+import { type HTMLProps, type PolymorphicProps, ark } from '../factory'
 import { useProgressContext } from './use-progress-context'
 
 export interface ProgressCircleTrackBaseProps extends PolymorphicProps {}
 export interface ProgressCircleTrackProps
-  extends SVGProps<SVGCircleElement>,
+  extends HTMLProps<'circle'>,
     ProgressCircleTrackBaseProps {}
 
-export const ProgressCircleTrack = (props: ProgressCircleTrackProps) => {
-  const progress = useProgressContext()
-  const mergedProps = mergeProps(progress.getCircleTrackProps(), props)
+export const ProgressCircleTrack = forwardRef<SVGCircleElement, ProgressCircleTrackProps>(
+  (props, ref) => {
+    const progress = useProgressContext()
+    const mergedProps = mergeProps(progress.getCircleTrackProps(), props)
 
-  return <ark.circle {...mergedProps} />
-}
+    return <ark.circle ref={ref} {...mergedProps} />
+  },
+)
 
 ProgressCircleTrack.displayName = 'ProgressCircleTrack'