+++ packages/react/CHANGELOG.md
@@ -25,6 +25,7 @@ description: All notable changes will be documented in this file.
   - Fixed problem where the context menu trigger showed the magnifier and iOS context menu on long press.
 - **Editable:** Fixed bug where setting `activationMode` to `dblclick` unexpectedly cleared the input value.
 - **Checkbox:** Added the missing `group` role for the Checkbox group component.
+- **Progress:** Converted `Circle`, `CircleTrack`, and `CircleRange` components to use React's `forwardRef` for improved ref handling.
 
 ## [3.3.0] - 2024-06-12